+++ build.gradle
@@ -8,7 +8,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.google.gms:google-services:4.3.14'
-        classpath 'com.android.tools.build:gradle:7.3.1'
+        classpath 'com.android.tools.build:gradle:7.4.1'
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.10'
         classpath 'com.google.firebase:firebase-crashlytics-gradle:2.9.4'
         // NOTE: Do not place your application dependencies here; they belong
